Chore: Move `load-resource` and its dependencies to `utils`
