+++ packages/hint/src/lib/enums/index.ts
@@ -1,4 +1,5 @@
 export * from '@hint/utils/dist/src/types/category';
 export * from './error-status';
 export * from './hint-scope';
-export * from './resource-type';
+export * from '@hint/utils/dist/src/types/resource-type';
+export * from '@hint/utils/dist/src/types/resource-error-status';
