+++ packages/hint/src/lib/enums/resource-type.ts
@@ -1,8 +1,1 @@
-/** The type of resource */
-export enum ResourceType {
-    configuration = 'configuration',
-    connector = 'connector',
-    formatter = 'formatter',
-    parser = 'parser',
-    hint = 'hint'
-}
+export * from '@hint/utils/dist/src/types/resource-type';
