--- packages/hint/tests/lib/utils/fixtures/@example/webhint-configuration-example/index.json
@@ -1,7 +1,0 @@
-{
-    "hints": {
-        "@example/webhint-hint-example": "error",
-        "@example2/webhint-hint-example2": "error",
-        "webhint-hint-example3": "error"
-    }
-}
