--- packages/hint/tests/lib/utils/fixtures/@example/webhint-configuration-example/test_modules/@example/webhint-hint-example/index.js
@@ -1,3 +1,0 @@
-console.log('@example/webhint-hint-example');
-
-module.exports = { meta: { id: 'example' } };
