--- packages/hint/tests/lib/utils/fixtures/@example/webhint-configuration-example/test_modules/@example/webhint-hint-example/package.json
@@ -1,6 +1,0 @@
-{
-  "main": "index.js",
-  "name": "@example/webhint-hint-example",
-  "private": true,
-  "version": "1.0.0"
-}
