--- packages/hint/tests/lib/utils/fixtures/@example/webhint-configuration-example/test_modules/@example2/webhint-hint-example2/index.js
@@ -1,3 +1,0 @@
-console.log('@example2/webhint-hint-example2');
-
-module.exports = { meta: { id: 'example2' } };
