--- packages/hint/tests/lib/utils/fixtures/@example/webhint-configuration-example/test_modules/@example2/webhint-hint-example2/package.json
@@ -1,6 +1,0 @@
-{
-  "main": "index.js",
-  "name": "@example2/webhint-hint-example2",
-  "private": true,
-  "version": "1.0.0"
-}
