--- packages/hint/tests/lib/utils/fixtures/@example/webhint-configuration-example/test_modules/webhint-hint-example3/index.js
@@ -1,3 +1,0 @@
-console.log('webhint-hint-example3');
-
-module.exports = { meta: { id: 'example3' } };
