--- packages/hint/tests/lib/utils/fixtures/@example/webhint-configuration-example/test_modules/webhint-hint-example3/package.json
@@ -1,6 +1,0 @@
-{
-  "main": "index.js",
-  "name": "webhint-hint-example3",
-  "private": true,
-  "version": "1.0.0"
-}
