+++ packages/utils/package.json
@@ -17,6 +17,7 @@
     "eventemitter2": "^5.0.1",
     "file-type": "^12.1.0",
     "file-url": "^3.0.0",
+    "globby": "^10.0.1",
     "is-svg": "^4.2.0",
     "is-wsl": "^2.1.0",
     "jsonc-parser": "^2.1.0",
