+++ packages/utils/src/config/index.ts
@@ -1,1 +1,2 @@
 export { default as normalizeHints } from './normalize-hints';
+export * from './to-absolute-path';
