+++ packages/utils/src/packages/index.ts
@@ -2,3 +2,8 @@ export * from './find-node-modules-root';
 export * from './find-package-root';
 export * from './is-official';
 export * from './load-package';
+export * from './load-resource';
+export * from './has-multiple-resources';
+export * from './is-full-package-name';
+export * from './load-hint-package';
+export * from './require-package';
