+++ packages/utils/src/types/resource-error-status.ts
@@ -1,0 +1,6 @@
+export enum ResourceErrorStatus {
+    DependencyError = 'DependencyError',
+    NotCompatible = 'NotCompatible',
+    NotFound = 'NotFound',
+    Unknown = 'Unknown'
+}
