+++ packages/utils/src/types/resource-type.ts
@@ -1,0 +1,8 @@
+/** The type of resource */
+export enum ResourceType {
+    configuration = 'configuration',
+    connector = 'connector',
+    formatter = 'formatter',
+    parser = 'parser',
+    hint = 'hint'
+}
